Updating atomicity logic & test cases.
