+++ docker-compose.yml
@@ -3,4 +3,4 @@ app:
   image: ibmcom/swift-ubuntu:3.1.1
   volumes:
     - .:/root/CircuitBreaker
-  command: bash -c "cd /root/CircuitBreaker && rm -rf Packages && swift package clean && swift build && swift test"
+  command: bash -c "cd /root/CircuitBreaker && swift package clean && swift build && swift test"
